Refactor data handling utilities into a separate module
Moved dataset classes, sampler creation, and dataloader creation logic from `utils.py` to a new `data_handlers.py` module. This improves code modularity and readability by isolating data handling responsibilities from other utilities. Updated imports across the codebase accordingly.
